feat(mock): add comprehensive mock logger for public testing use
Create a new handlers/mock package that provides Logger and Recorder types
for testing slog-based applications. This moves and enhances the testing
capabilities from internal/testing to make them publicly available.

Key features:
- Thread-safe Logger and Recorder implementations with comprehensive nil checks
- Complete slog.Logger interface implementation with immutable behavior
- Optimized Message.String() using strings.Builder with proper error handling
- Comprehensive test suite following project guidelines with factory functions
- Full documentation with usage examples and migration guidance

Implementation details:
- Logger captures all log levels, fields, stack traces, and print methods
- Recorder provides thread-safe message storage with copy semantics
- All methods handle nil receivers gracefully to prevent panics
- Field chains are properly copied to maintain immutability
- Message formatting sorts fields consistently for predictable output

Backward compatibility:
- internal/testing now provides deprecated type aliases
- Existing code continues to work without modification
- Documentation encourages migration to public API
- Clear migration path with examples provided

Documentation:
- handlers/mock/README.md with comprehensive usage examples
- Updated internal/testing/README.md with migration guidance
- Root README.md includes mock handler in utility handlers section
- All examples show both old and new usage patterns

This enables slog users to test their own logging code by capturing and
verifying log messages programmatically, previously only available to
internal slog handler development.
